gateway/requirements.txt
nameko==2.14.1
amqp==5.2.0
PyJWT==2.8.0